@@ -4,23 +4,21 @@
 namespace HomeCEU\DTS\Api\Render;
 
 
-use HomeCEU\DTS\Api\ApiHelper;
 use HomeCEU\DTS\Persistence\CompiledTemplatePersistence;
 use HomeCEU\DTS\Persistence\HotRenderPersistence;
 use HomeCEU\DTS\Persistence\TemplatePersistence;
 use HomeCEU\DTS\Render\CompilationException;
 use HomeCEU\DTS\Repository\HotRenderRepository;
 use HomeCEU\DTS\Repository\TemplateRepository;
+use HomeCEU\DTS\UseCase\Exception\InvalidHotRenderRequestException;
 use HomeCEU\DTS\UseCase\Render\AddHotRender as AddHotRenderUseCase;
 use HomeCEU\DTS\UseCase\Render\AddHotRenderRequest;
-use HomeCEU\DTS\UseCase\Render\InvalidHotRenderRequestException;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\RequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 
 class AddHotRender {
-  /** @var \HomeCEU\DTS\UseCase\Render\AddHotRender */
-  private $useCase;
+  private AddHotRenderUseCase $useCase;
 
   public function __construct(ContainerInterface $container) {
     $conn = $container->dbConnection;
