@@ -4,7 +4,6 @@
 namespace HomeCEU\DTS\Api\Template;
 
 
-use Exception;
 use HomeCEU\DTS\Api\ResponseHelper;
 use HomeCEU\DTS\Entity\Template;
 use HomeCEU\DTS\Persistence\CompiledTemplatePersistence;
@@ -16,8 +15,7 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\Http\Response;
 
 class ListVersions {
-
-  private $useCase;
+  private TemplateVersionList $useCase;
 
   public function __construct(ContainerInterface $diContainer) {
     $db = $diContainer->dbConnection;
