@@ -7,19 +7,9 @@ namespace HomeCEU\DTS\Entity;
 use HomeCEU\DTS\AbstractEntity;
 
 class DocData extends AbstractEntity {
-  public $dataId;
-  public $docType;
-  public $dataKey;
-  public $createdAt;
-  public $data;
-
-  protected static function keys(): array {
-    return [
-        'dataId',
-        'docType',
-        'dataKey',
-        'createdAt',
-        'data'
-    ];
-  }
+  public string $dataId;
+  public string $docType;
+  public string $dataKey;
+  public \DateTime $createdAt;
+  public array $data;
 }
