@@ -7,17 +7,8 @@ namespace HomeCEU\DTS\Entity;
 use HomeCEU\DTS\AbstractEntity;
 
 class HotRenderRequest extends AbstractEntity {
-  public $requestId;
-  public $template;
-  public $data;
-  public $createdAt;
-
-  protected static function keys(): array {
-    return [
-        'requestId',
-        'template',
-        'data',
-        'createdAt',
-    ];
-  }
+  public string $requestId;
+  public string $template;
+  public array $data;
+  public \DateTime $createdAt;
 }
