@@ -7,13 +7,6 @@ namespace HomeCEU\DTS\UseCase\Render;
 use HomeCEU\DTS\AbstractEntity;
 
 class HotRenderRequest extends AbstractEntity {
-  public $requestId;
-  public $format;
-
-  protected static function keys(): array {
-    return [
-        'requestId',
-        'format'
-    ];
-  }
+  public string $requestId;
+  public string $format;
 }
