@@ -7,13 +7,6 @@ namespace HomeCEU\DTS\UseCase\Render;
 use HomeCEU\DTS\AbstractEntity;
 
 class RenderResponse extends AbstractEntity {
-  public $path;
-  public $contentType;
-
-  protected static function keys(): array {
-    return [
-        'path',
-        'contentType'
-    ];
-  }
+  public string $path;
+  public string $contentType;
 }
