@@ -3,24 +3,21 @@
 
 namespace HomeCEU\DTS\Api;
 
-use HomeCEU\DTS\Db;
-use HomeCEU\DTS\Render\TemplateCompiler;
-use Monolog\Logger;
+
 use Psr\Container\ContainerInterface;
 use Slim\Container;
 
-/**
- * Class DiContainer
- * @property Db\Config dbConfig
- * @property Db\Connection dbConnection
- * @property TemplateCompiler templateCompiler
- * @property Logger logger
- */
 class DiContainer extends Container implements ContainerInterface {
   public function __construct(array $values = []) {
-    if ($values == []) {
-      $values = include __DIR__."/services.php";
+    $values = array_merge($values, include __DIR__."/services.php");
+    parent::__construct($this->loadEnvironmentalServices($values));
+  }
+
+  protected function loadEnvironmentalServices(array $values): array {
+    $env = strtolower(getenv('APP_ENV'));
+    if (is_file(__DIR__ . "/services.{$env}.php")) {
+      $values = array_merge($values, include __DIR__ . "/services.{$env}.php");
     }
-    parent::__construct($values);
+    return $values;
   }
 }
