@@ -5,9 +5,9 @@ namespace HomeCEU\DTS;
 
 
 interface Persistence {
-  public function generateId();
-  public function persist($data);
-  public function retrieve($id, array $cols=['*']);
-  public function find(array $filter, array $cols=['*']);
+  public function generateId(): string;
+  public function persist($data): string;
+  public function retrieve($id, array $cols=['*']): array;
+  public function find(array $filter, array $cols=['*']): array;
   public function delete($id);
 }
