@@ -7,8 +7,7 @@ namespace HomeCEU\DTS\Persistence\InMemory;
 use HomeCEU\DTS\Persistence\InMemory;
 
 class DocDataPersistence extends InMemory {
-
-  public function getTable() {
+  public function getTable(): string {
     return 'docdata';
   }
 
