@@ -4,7 +4,7 @@
 namespace HomeCEU\Tests\Api;
 
 
-class StatusTest extends TestCase {
+class StatusTest extends ApiTestCase {
   public function testStatusReturnsOk(): void {
     $response = $this->get('/api/v1/status');
     $this->assertStatus(200, $response);
