@@ -4,9 +4,10 @@
 namespace HomeCEU\Tests\Api\Template;
 
 
+use HomeCEU\Tests\Api\ApiTestCase;
 use PHPUnit\Framework\Assert;
 
-class TestCase extends \HomeCEU\Tests\Api\TestCase {
+class TestCase extends ApiTestCase {
   protected function httpGetTemplatesFromUri($uri) {
     $response = $this->get($uri);
     $responseData = json_decode($response->getBody(), true);
