@@ -8,7 +8,7 @@ use HomeCEU\DTS\Persistence;
 use HomeCEU\DTS\Render\TemplateCompiler;
 use HomeCEU\DTS\Repository\DocDataRepository;
 use HomeCEU\DTS\Repository\TemplateRepository;
-use HomeCEU\DTS\UseCase\Render\InvalidRenderRequestException;
+use HomeCEU\DTS\UseCase\Exception\InvalidRenderRequestException;
 use HomeCEU\DTS\UseCase\Render\Render;
 use HomeCEU\DTS\UseCase\Render\RenderFormat;
 use HomeCEU\DTS\UseCase\Render\RenderRequest;
